requirements/local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=local.txt local.in
#
aspy.refactor-imports==1.1.0
aspy.yaml==1.3.0          # via pre-commit
atomicwrites==1.3.0
attrs==19.3.0
cached-property==1.5.1
certifi==2019.9.11
cfgv==2.0.1               # via pre-commit
chardet==3.0.4
click==7.0                # via pip-tools
django-admin-ordering==0.11.0
django-anymail[mailgun]==7.0.0
django-ckeditor==5.7.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.3
django-js-asset==1.2.2
django-nested-admin==3.2.4
django==2.2.6
easy-thumbnails==2.6
entrypoints==0.3
factory-boy==2.12.0
factory-djoy==2.1.1
faker==2.0.3
flake8==3.7.8
identify==1.4.7           # via pre-commit
idna==2.8
importlib-metadata==0.23
importlib-resources==1.0.2  # via pre-commit
isort==4.3.21
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3            # via pre-commit
packaging==19.2
pillow==6.2.0
pip-tools==4.2.0
pluggy==0.13.0
pre-commit==1.18.3
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest-django==3.5.1
pytest==5.2.1
python-dateutil==2.8.0
python-monkey-business==1.0.0
pytz==2019.3
pyyaml==5.1.2             # via aspy.yaml, pre-commit
requests==2.22.0
seed-isort-config==1.9.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0              # via pre-commit
twitter==1.18.0
urllib3==1.25.6
virtualenv==16.7.6        # via pre-commit
wcwidth==0.1.7
yapf==0.28.0
zipp==0.6.0
